refactor: split borg models from execution
